Add files via upload
Added button group in script.html containing Alert Button, Hover Button and Increment Button.
Added button group in script.html containing Alert Button, Hover Button and Increment Button.
Added div below the button group with the message "Chicken is". The for loop will append the word "great" for x amount of times to the end of the message.
Changes:
Added button group in script.html containing Alert Button, Hover Button and Increment Button.
Added div below the button group with the message "Chicken is". The for loop will append the word "great" for x amount of times to the end of the message.